+++ keyboards/dz60/keymaps/jdelkins/rules.mk
@@ -1,7 +1,7 @@
 LEADER_ENABLE = yes
-MOUSEKEY_ENABLE = yes
+MOUSEKEY_ENABLE = no
 CONSOLE_ENABLE = no
 NKRO_ENABLE = no
 TAP_DANCE_ENABLE = yes
-LTO_ENABLE = yes
 AUTO_SHIFT_ENABLE = yes
+LTO_ENABLE = yes
