--- keyboards/dz60/keymaps/jdelkins_ss/rules.mk
@@ -1,8 +1,0 @@
-LEADER_ENABLE = yes
-MOUSEKEY_ENABLE = yes
-CONSOLE_ENABLE = no
-NKRO_ENABLE = no
-TAP_DANCE_ENABLE = yes
-AUTO_SHIFT_ENABLE = yes
-LTO_ENABLE = yes
-USER_NAME := jdelkins
