+++ users/jdelkins/rules.mk
@@ -1,10 +1,15 @@
 SRC += jdelkins.c
 
-ifeq ($(strip $(NO_SECRETS)), yes)
-    OPT_DEFS += -DNO_SECRETS
+ifneq ($(strip $(NO_SECRETS)),)
+	OPT_DEFS += -DNO_SECRETS
 endif
+
+ifeq ($(shell test -f users/jdelkins/secrets.h.gpg && echo yes || echo no),yes)
+users/jdelkins/jdelkins.c: users/jdelkins/secrets.h
+	touch $@
 
 users/jdelkins/secrets.h: users/jdelkins/secrets.h.gpg
 	gpg -d $< >$@
+endif
 
 BOOTMAGIC_ENABLE = no
